+++ js/data.js
@@ -1,39 +1,55 @@
-/* eslint-disable no-alert */
-const getRandomInteger = (min = 1, max = 5) => {
-  if (min >= max || min < 0) {
-    alert('Неверный ввод данных!');
-  }
-  const integerNumber = Math.floor(min + Math.random() * (max + 1 - min));
-  return +integerNumber;
-};
+import {ARRAY_USERS_COUNT, PLACE_AND_TIME, FACILITIES} from './constants.js';
+import {getRandomInteger, getNumberImg, getRandomArray, longitude, latitude} from './until.js';
 
-const getRandomFloat = (min = 1, max = 5, digits = 1) => {
-  if (min >= max || min < 0) {
-    alert('Неверный ввод данных!');
-  }
-  const floatNumber = (min + Math.random() * (max - min)).toFixed(digits);
-  return +floatNumber;
-};
-
-const getNumberImg = (arrayLength = 10) => {
-  const numbersImg = [...Array(Math.abs(arrayLength)).keys()];
-  const addZero = numbersImg.map((currentValue) => {
-    currentValue = (currentValue + 1).toString();
-    return currentValue.padStart(2, '0');
+const getAvatar = () => {
+  const addressImg = getNumberImg().map((currentValue) => {
+    currentValue = `img/avatars/user${currentValue}.png`;
+    return currentValue;
   });
-  return addZero;
+  return addressImg;
 };
 
-const getRandomArray = (value) => {
-  const arrayOfArbitraryLength = Array(getRandomInteger(0, value.length));
-  for (let i = 0; i < arrayOfArbitraryLength.length; i++) {
-    const randomNumber = getRandomInteger(0, arrayOfArbitraryLength.length);
-    if (!arrayOfArbitraryLength.includes(randomNumber)) {
-      arrayOfArbitraryLength[i] = randomNumber;
-    }
+const getOneOfMany = (min, max, value) => {
+  const array = value;
+  return array[getRandomInteger(min, max) - 1];
+};
+
+const getArrayRandomValues = (value) => {
+  const randomArray = getRandomArray(value);
+  const arrayRandomValues = [];
+  for (let i = 0; i < randomArray.length; i++) {
+    arrayRandomValues.push(value[randomArray[i]]);
   }
-  const filteredArray = arrayOfArbitraryLength.filter((currentValue) => currentValue !== undefined);
+  const filteredArray = arrayRandomValues.filter((currentValue) => currentValue !== undefined);
   return filteredArray;
 };
 
-export{getRandomInteger, getRandomFloat, getNumberImg, getRandomArray};
+const createUser = () => ({
+  author: {
+    avatar: getOneOfMany(1, 10, getAvatar()),
+  },
+  offer: {
+    title: 'Какой-то заголовок',
+    address: {
+      lat: latitude,
+      lng: longitude,
+    },
+    price: getRandomInteger(1000, 10000),
+    type: getOneOfMany(1, 5, PLACE_AND_TIME.apartments),
+    rooms: getRandomInteger(1, 6),
+    guests: getRandomInteger(1, 20),
+    checkin: getOneOfMany(1, 3, PLACE_AND_TIME.time),
+    checkout: getOneOfMany(1, 3, PLACE_AND_TIME.time),
+    features: getArrayRandomValues(FACILITIES.additions),
+    description: 'Какое-то описание',
+    photos: getArrayRandomValues(FACILITIES.interior),
+  },
+  location: {
+    lat: latitude,
+    lng: longitude,
+  }
+});
+
+const createArrayUsers = () => Array.from({length: ARRAY_USERS_COUNT}, createUser);
+
+export {createArrayUsers};
