+++ js/main.js
@@ -1,1 +1,5 @@
-import './util.js';
+/* eslint-disable no-console */
+import {createArrayUsers} from './data.js';
+
+console.log(createArrayUsers());
+
